+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Are the clumps normal".
 - Tests for "Bulk email generator".
 - Tests for "Snail run".
+- Tests for "Sweet spot".
 
 ## [3.8.0] - 2022-04-30
 ### Added